♻️ refactor: separate common and sync slice
