+++ src/store/user/slices/common/action.ts
@@ -8,15 +8,12 @@ import { UserConfig, userService } from '@/services/user';
 import type { UserStore } from '@/store/user';
 import type { GlobalServerConfig } from '@/types/serverConfig';
 import type { GlobalSettings } from '@/types/settings';
-import { OnSyncEvent, PeerSyncStatus } from '@/types/sync';
 import { switchLang } from '@/utils/client/switchLang';
 import { merge } from '@/utils/merge';
-import { browserInfo } from '@/utils/platform';
 import { setNamespace } from '@/utils/storeDebug';
 
 import { preferenceSelectors } from '../preference/selectors';
-import { settingsSelectors, syncSettingsSelectors } from '../settings/selectors';
-import { userProfileSelectors } from './selectors';
+import { settingsSelectors } from '../settings/selectors';
 
 const n = setNamespace('common');
 
@@ -24,16 +21,9 @@ const n = setNamespace('common');
  * 设置操作
  */
 export interface CommonAction {
-  refreshConnection: (onEvent: OnSyncEvent) => Promise<void>;
   refreshUserConfig: () => Promise<void>;
-  triggerEnableSync: (userId: string, onEvent: OnSyncEvent) => Promise<boolean>;
   updateAvatar: (avatar: string) => Promise<void>;
   useCheckTrace: (shouldFetch: boolean) => SWRResponse;
-  useEnabledSync: (
-    userEnableSync: boolean,
-    userId: string | undefined,
-    onEvent: OnSyncEvent,
-  ) => SWRResponse;
   useFetchServerConfig: () => SWRResponse;
   useFetchUserConfig: (initServer: boolean) => SWRResponse<UserConfig | undefined>;
 }
@@ -46,14 +36,6 @@ export const createCommonSlice: StateCreator<
   [],
   CommonAction
 > = (set, get) => ({
-  refreshConnection: async (onEvent) => {
-    const userId = userProfileSelectors.userId(get());
-
-    if (!userId) return;
-
-    await get().triggerEnableSync(userId, onEvent);
-  },
-
   refreshUserConfig: async () => {
     await mutate([USER_CONFIG_FETCH_KEY, true]);
 
@@ -61,38 +43,6 @@ export const createCommonSlice: StateCreator<
     get().refreshModelProviderList();
   },
 
-  triggerEnableSync: async (userId: string, onEvent: OnSyncEvent) => {
-    // double-check the sync ability
-    // if there is no channelName, don't start sync
-    const sync = syncSettingsSelectors.webrtcConfig(get());
-    if (!sync.channelName) return false;
-
-    const name = syncSettingsSelectors.deviceName(get());
-
-    const defaultUserName = `My ${browserInfo.browser} (${browserInfo.os})`;
-
-    set({ syncStatus: PeerSyncStatus.Connecting });
-    return globalService.enabledSync({
-      channel: {
-        name: sync.channelName,
-        password: sync.channelPassword,
-      },
-      onAwarenessChange(state) {
-        set({ syncAwareness: state });
-      },
-      onSyncEvent: onEvent,
-      onSyncStatusChange: (status) => {
-        set({ syncStatus: status });
-      },
-      signaling: sync.signaling,
-      user: {
-        id: userId,
-        // if user don't set the name, use default name
-        name: name || defaultUserName,
-        ...browserInfo,
-      },
-    });
-  },
   updateAvatar: async (avatar) => {
     await userService.updateAvatar(avatar);
     await get().refreshUserConfig();
@@ -115,25 +65,6 @@ export const createCommonSlice: StateCreator<
       },
     ),
 
-  useEnabledSync: (userEnableSync, userId, onEvent) =>
-    useSWR<boolean>(
-      ['enableSync', userEnableSync, userId],
-      async () => {
-        // if user don't enable sync or no userId ,don't start sync
-        if (!userId) return false;
-
-        // if user don't enable sync, stop sync
-        if (!userEnableSync) return globalService.disableSync();
-
-        return get().triggerEnableSync(userId, onEvent);
-      },
-      {
-        onSuccess: (syncEnabled) => {
-          set({ syncEnabled }, false, n('useEnabledSync'));
-        },
-        revalidateOnFocus: false,
-      },
-    ),
   useFetchServerConfig: () =>
     useSWR<GlobalServerConfig>('fetchGlobalConfig', globalService.getGlobalConfig, {
       onSuccess: (data) => {
