+++ src/store/user/slices/sync/initialState.ts
@@ -1,17 +1,12 @@
 import { PeerSyncStatus, SyncAwarenessState } from '@/types/sync';
 
-export interface Guide {
-  // Topic 引导
-  topic?: boolean;
-}
-
-export interface UserCommonState {
+export interface UserSyncState {
   syncAwareness: SyncAwarenessState[];
   syncEnabled: boolean;
   syncStatus: PeerSyncStatus;
 }
 
-export const initialCommonState: UserCommonState = {
+export const initialSyncState: UserSyncState = {
   syncAwareness: [],
   syncEnabled: false,
   syncStatus: PeerSyncStatus.Disabled,
